+++ app/models/tutorial.model.js
@@ -1,0 +1,18 @@
+// automatically added by vs-code
+const { Sequelize } = require(".");
+
+module.exports = (sequelize, Sequelize) => {
+  const Tutorial = sequelize.define("tutorial", {
+    title: {
+      type: Sequelize.STRING
+    },
+    description: {
+      type: Sequelize.STRING
+    },
+    published: {
+      type: Sequelize.BOOLEAN
+    }
+  });
+
+  return Tutorial;
+};